requirements.txt
discord.py==1.4.1
pylint==2.6.0
python-dotenv==0.14.0